requirements.txt
flask
python-dotenv
requests
supabase
schedule
python-telegram-bot
gunicorn